Model Development
The file include a Python code that shows the Model development of Laptop Price data using Single and Multiple Linear Regression as well as the Polynomial Regression. Then the results were compared  using Mean Squared Error (MSE) , R-squared and the Visualisations to evaluate the Models' fit to the actual data. Also, Pipeline is used to show how model development tasks can be simplified, automated, and streamlined.